Refactor SearchBar and SearchResults
