+++ client/src/components/shared/SearchBar/SearchResults/index.js
@@ -1,7 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
 
-import { toTitleCase } from "../../utils";
+import { toTitleCase } from "../../../../utils";
 
 const SearchResults = ({
   isOrigin,
@@ -14,33 +14,23 @@ const SearchResults = ({
     // Render results if available
     // and search value has more than 2 characters
     if (searchResults.length > 0 && searchValue.length > 2)
-      return searchResults.map((result) => {
-        const id = result["id"];
-        const namePriamry = result["labelPriamry"];
-        const nameSecondary = result["labelSecondary"];
-        const layer = toTitleCase(result["layer"]);
-
-        return (
-          <li onMouseDown={() => selectResult(result)} key={id}>
-            {namePriamry}
-            <span>
-              {nameSecondary} ({layer})
-            </span>
-          </li>
-        );
-      });
+      return searchResults.map((result) => (
+        <li onMouseDown={() => selectResult(result)} key={result["id"]}>
+          {result["labelPriamry"]}
+          <span>
+            {result["labelSecondary"]} ({toTitleCase(result["layer"])})
+          </span>
+        </li>
+      ));
   }
+
+  const key = isOrigin
+    ? "58bfeeaf-5785-4966-9037-b248397608a7-origin"
+    : "58bfeeaf-5785-4966-9037-b248397608a7-destination";
 
   return (
     <ul>
-      <li
-        onMouseDown={getGeolocation}
-        key={
-          isOrigin
-            ? "58bfeeaf-5785-4966-9037-b248397608a7-origin"
-            : "58bfeeaf-5785-4966-9037-b248397608a7-destination"
-        }
-      >
+      <li onMouseDown={getGeolocation} key={key}>
         Use Current Location
       </li>
       {renderResults()}
