+++ client/src/components/shared/TripIcon/index.js
